"Integrated JaCoCo for code coverage reports
Jacoco plugin has been added to the build.gradle file in this commit. This was done to incorporate code coverage statistics in the project. The version used is 0.8.9 and test reports will be generated upon running tests. HTML report output will be saved in 'jacocoHtml' directory in the build directory of every main folder"