Add tests for align object getValue method. Still need to add comments though
Adding three new tests in the AlignTests class. These tests cover the getValue method of the align object when this is set to right, left, and center align. The tests ensure that the returned value is correctly calculated from the product of the width and the portion. Increased the Align class from 87% method percentage to 100% and line 95% to 100%.